requirements.txt
# API Integrations
requests>=2.28.1
anthropic>=0.5.0
google-api-python-client>=2.79.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0

# Data Processing
pandas>=1.5.3
beautifulsoup4>=4.11.1
markdownify>=0.11.6
html2text>=2020.1.16

# Async Support
aiohttp>=3.8.4
asyncio>=3.4.3

# Scheduling
apscheduler>=3.10.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pyyaml>=6.0

# Optional Development Tools
pytest>=7.3.1
black>=23.3.0
flake8>=6.0.0
